refactor: move weather related behavior into different layers
