+++ Doppler.HelloMicroservice/Weather/DataService.cs
@@ -1,0 +1,10 @@
+namespace Doppler.HelloMicroservice.Weather
+{
+    public class DataService
+    {
+        public string[] GetData() => new[]
+        {
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
+    }
+}
